expense/requirements.txt
asn1crypto==0.22.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
contextlib2==0.5.5
cryptography==1.9
Django==1.11.2
django-anymail==0.11.1
django-bootstrap3==9.0.0
django-crontab==0.7.1
django-excel==0.0.9
django-extensions==1.8.1
django-filter==1.0.4
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
drf-extensions==0.3.1
gunicorn==19.7.1
idna==2.5
lml==0.0.1
phonenumbers==8.6.0
pycparser==2.18
pyexcel==0.5.3
pyexcel-io==0.4.4
pyexcel-webio==0.1.2
pyexcel-xls==0.4.0
PyJWT==1.5.2
pytz==2017.2
raven==6.1.0
requests==2.18.1
six==1.10.0
texttable==0.9.1
urllib3==1.21.1
uWSGI==2.0.15
whitenoise==3.3.0
xlrd==1.0.0
xlwt==1.2.0